+++ lib/vagrant/driver/virtualbox_base.rb
@@ -133,9 +133,8 @@ module Vagrant
       # Imports the VM from an OVF file.
       #
       # @param [String] ovf Path to the OVF file.
-      # @param [String] name Name of the VM.
       # @return [String] UUID of the imported VM.
-      def import(ovf, name)
+      def import(ovf)
       end
 
       # Returns a list of forwarded ports for a VM.